+++ applications/ALEapplication/tests/mpi_test_laplacian_mesh_motion_3d.py
@@ -3,35 +3,19 @@ import KratosMultiphysics
 import KratosMultiphysics.ALEApplication
 import KratosMultiphysics.KratosUnittest as KratosUnittest
 import KratosMultiphysics.kratos_utilities as kratos_utils
-import ale_analysis
+import mesh_moving_test_case
 
-class ControlledExecutionScope:
-    def __init__(self, scope):
-        self.currentPath = os.getcwd()
-        self.scope = scope
-
-    def __enter__(self):
-        os.chdir(self.scope)
-
-    def __exit__(self, type, value, traceback):
-        os.chdir(self.currentPath)
-
-class TestCase(KratosUnittest.TestCase):
-
-    def createTest(self, parameter_file_name):
-        with open(parameter_file_name + '_parameters.json', 'r') as parameter_file:
-            self.project_parameters = KratosMultiphysics.Parameters(parameter_file.read())
-
+class TestCase(mesh_moving_test_case.MeshMovingTestCase):
     def test_Rectangle_3D8N(self):
-        with ControlledExecutionScope(os.path.dirname(os.path.realpath(__file__))):
+        with mesh_moving_test_case.ControlledExecutionScope(os.path.dirname(os.path.realpath(__file__))):
             self.createTest('test_laplacian_mesh_motion_3d/mpi_rectangle_3D8N_test')
-            ale_analysis.ALEAnalysis(self.project_parameters).Run()
+            self.runTest()
             kratos_utils.DeleteFileIfExisting("./test_mdpa_files/rectangle_3D8N_test.time")
 
     def test_Rectangle_3D4N(self):
-        with ControlledExecutionScope(os.path.dirname(os.path.realpath(__file__))):
+        with mesh_moving_test_case.ControlledExecutionScope(os.path.dirname(os.path.realpath(__file__))):
             self.createTest('test_laplacian_mesh_motion_3d/mpi_rectangle_3D4N_test')
-            ale_analysis.ALEAnalysis(self.project_parameters).Run()
+            self.runTest()
             kratos_utils.DeleteFileIfExisting("./test_mdpa_files/rectangle_3D4N_test.time")
